Query for connection ID after connecting
Motivation: As observed on some installations (e.g. Azure hosted MySQL
5.6.39), the connection ID returned in the handshake packet might be
different from the actual connection ID.

This issue can be reproduced easily in the console:

  > mysql

  Welcome to the MySQL monitor.  Commands end with ; or \g.
  Your MySQL connection id is 64937
  Server version: 5.6.39.0 MySQL Community Server (GPL)

  (...)

  mysql> select connection_id();
  +-----------------+
  | connection_id() |
  +-----------------+
  |           15567 |
  +-----------------+
  1 row in set (0.04 sec)

This happens on an Azure installation while it does not happen e.g. on
a local installation of MySQL 5.6.

There is a related issue: https://bugs.mysql.com/bug.php?id=44167
It also suggests that the connection ID from the handshake might not be
reliable, although the case described here is different from that one.

As the connection ID is used to kill queries, it needs to be correct or
the DB worker process would crash.